backend/pip_requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
python-dotenv
aiofiles

passlib[bcrypt]
